refactor cli and dump more metadata into logs
